Send help
My brain is melting.
I switched the purposes of TA2 and TA3 because the encoders are hardwired to TA3 pins, meaning that I want to use them to trigger interrupts somehow. I'm only marginally sure that I've set up the pins correctly to do this and can't find a definition anywhere.