Final Updates
I wasn't able to get things working. I realized after probably 12 hours of work that... PeriodMeasure already measures the speed of the motors, meaning that all of my work is pointless. I have no clue how PeriodMeasure works, which makes matters worse. Basically I've run out of time to work on this so this will be the final update,